test(sdk): separate runtime Swagger cases
Close #1590
